Fix error by filtering out deleted files
When a .py file that was deployed to Lambda was deleted from the repository, this Action would detect the difference for that file (even though it was deleted) and try to package it up to send to Lambda. But because the file is deleted, it couldn't find the file, causing an error. This commit addresses this by filtering out deleted items from the git diff.