Restructure server code
Restructuring the server code to make it easier to understand, maintain,
and deploy. There is a "common" module with common functionality, like DB
and Twilio. There is now a "functions" folder containing all lambda
functions, and a new draft for a CI/CD workflow. Will be implementing
that in the coming commits